+++ source/tessellation/static_voronoi_mesh.cpp
@@ -449,22 +449,22 @@ void StaticVoronoiMesh::build_v()
 		to_check=Tri.get_facet(i);
 		for(j=0;j<3;++j)
 		{
-			if(to_check.neighbors[static_cast<size_t>(j)]==Tri.get_last_loc())
+		  if(to_check.neighbors[static_cast<size_t>(j)]==static_cast<size_t>(Tri.get_last_loc()))
 				continue;
-			if(to_check.neighbors[static_cast<size_t>(j)]<i)
+		  if(to_check.neighbors[static_cast<size_t>(j)]<static_cast<size_t>(i))
 				continue;
 			center_temp=centers[static_cast<size_t>(to_check.neighbors[static_cast<size_t>(j)])];
 			{
 				edge_temp.vertices.first = center;
 				edge_temp.vertices.second = center_temp;
-				edge_temp.neighbors.first = to_check.vertices[static_cast<size_t>(j)];
-				edge_temp.neighbors.second = to_check.vertices[static_cast<size_t>(j+1)%3];
+				edge_temp.neighbors.first = static_cast<int>(to_check.vertices[static_cast<size_t>(j)]);
+				edge_temp.neighbors.second = static_cast<int>(to_check.vertices[static_cast<size_t>(j+1)%3]);
 
 				if(legal_edge(&edge_temp))
 				{
 					// I added a change here, if edge has zero length I don't add it.
 					if(edge_temp.GetLength()>eps*sqrt(Tri.GetFacetRadius(i)*
-						Tri.GetFacetRadius(to_check.neighbors[static_cast<size_t>(j)])))
+									  Tri.GetFacetRadius(static_cast<int>(to_check.neighbors[static_cast<size_t>(j)]))))
 					{
 						{
 							if(edge_temp.neighbors.first<Tri.GetOriginalLength())
